requirements.txt
python_dotenv==1.0.0
#from dotenv import load_dotenv
#load_dotenv()
openai==0.27.8 